app/requirements-dev.txt
-r requirements.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Code quality
flake8==6.1.0
pylint==3.0.3
black==23.12.1
bandit==1.7.6

# Security
safety==2.3.5
